Current + calibrations and refactor

+++ main/modules/base/types.h
@@ -1,4 +1,6 @@
 #include <stdint.h>
+#include <stddef.h>
+// #include <stdlib.h>
 
 
 typedef uint32_t Seconds;
@@ -7,7 +9,7 @@ typedef uint32_t Milliseconds;
 typedef int Volt;
 typedef int Millivolt;
 
-typedef uint32_t Amper;
+typedef float Amper;
 typedef uint32_t Milliamper;
 
 typedef uint32_t Herz;
